+++ packages/components/ui/rollup.config.js
@@ -12,7 +12,7 @@ import * as _ from 'lamb';
 import {makeBanner, renameToMinJs} from '@svizzle/dev';
 
 import pkg from './package.json';
-import rollupIconsInput from './src/rollup/rollupIconsInput.json';
+import rollupIconsInputFeather from './rollup/rollupIconsInputFeather.json';
 
 const analyzer = analyze({
 	limit: 15,
@@ -25,7 +25,8 @@ const dir = 'dist';
 const external = pkg.peerDependencies && Object.keys(pkg.peerDependencies) || [];
 
 const input = {
-	...rollupIconsInput,
+	...rollupIconsInputFeather,
+	A11yPerson: 'src/icons/svizzle/A11yPerson.svelte',
 	defaults: 'src/defaults.js',
 	ExternalLink: 'src/ExternalLink.svelte',
 	Icon: 'src/icons/Icon.svelte',