+++ packages/components/ui/src/icons/svizzle/index.js
@@ -1,0 +1,1 @@
+export {default as A11yPerson} from './A11yPerson.svelte';